commentary fixes
